+++ src/main/java/org/saltyrtc/client/signaling/ResponderSignaling.java
@@ -11,6 +11,7 @@ package org.saltyrtc.client.signaling;
 import com.neilalexander.jnacl.NaCl;
 
 import org.saltyrtc.client.SaltyRTC;
+import org.saltyrtc.client.annotations.NonNull;
 import org.saltyrtc.client.annotations.Nullable;
 import org.saltyrtc.client.cookie.Cookie;
 import org.saltyrtc.client.exceptions.ConnectionException;
@@ -50,9 +51,14 @@ public class ResponderSignaling extends Signaling {
         return org.slf4j.LoggerFactory.getLogger("SaltyRTC.RSignaling");
     }
 
+    @NonNull
     private final Initiator initiator;
-    private final AuthToken authToken;
+    @Nullable
+    private AuthToken authToken = null;
 
+    /**
+     * Create an instance without a trusted key.
+     */
     public ResponderSignaling(SaltyRTC saltyRTC, String host, int port,
                               KeyStore permanentKey, SSLContext sslContext,
                               byte[] initiatorPublicKey, byte[] authToken)
@@ -61,6 +67,20 @@ public class ResponderSignaling extends Signaling {
         this.role = SignalingRole.Responder;
         this.initiator = new Initiator(initiatorPublicKey);
         this.authToken = new AuthToken(authToken);
+    }
+
+    /**
+     * Create an instance with a trusted key.
+     */
+    public ResponderSignaling(SaltyRTC saltyRTC, String host, int port,
+                              KeyStore permanentKey, SSLContext sslContext,
+                              byte[] initiatorTrustedKey)
+            throws java.security.InvalidKeyException {
+        super(saltyRTC, host, port, permanentKey, sslContext, initiatorTrustedKey);
+        this.role = SignalingRole.Responder;
+        this.initiator = new Initiator(initiatorTrustedKey);
+        // If we trust the initiator, don't send a token message
+        this.initiator.handshakeState = InitiatorHandshakeState.TOKEN_SENT;
     }
 
     /**
@@ -157,7 +177,10 @@ public class ResponderSignaling extends Signaling {
     @Override
     protected void initPeerHandshake() throws ProtocolException, ConnectionException {
         if (this.initiator.isConnected()) {
-            this.sendToken();
+            // Only send token if we don't trust the initiator
+            if (this.peerTrustedKey == null) {
+                this.sendToken();
+            }
         }
     }
 
@@ -259,8 +282,10 @@ public class ResponderSignaling extends Signaling {
      * A new responder wants to connect.
      */
     protected void handleNewInitiator(NewInitiator msg) throws ProtocolException, ConnectionException {
-        // Initiator changed, send token
-        this.sendToken();
+        // Initiator changed, send token message if we don't trust the initiator
+        if (this.peerTrustedKey == null) {
+            this.sendToken();
+        }
     }
 
     /**